JL8-103 refactor : Header 컴포넌트 좌, 우 리팩토링
